% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bind_data}
\alias{bind_data}
\title{Bind data to project}
\usage{
bind_data(project, data_coverage, data_counts, data_sample, data_locus,
  name = NULL, check_delete_output = TRUE)
}
\arguments{
\item{project}{a \code{mipmapper_project}, as produced by the function 
\code{mipmapper_project()}}

\item{data_coverage}{a matrix giving the number of reads for each
sample-locus combination (see details)}

\item{data_counts}{a matrix giving the number of reads matched to the Ref
allele for each sample-locus combination (see details)}

\item{data_sample}{a dataframe giving details of each sample (see details)}

\item{data_locus}{a dataframe giving details of each locus (see details)}

\item{name}{optional name of the data set to aid in record keeping}

\item{check_delete_output}{whether to prompt the user before overwriting 
existing data/outputs}
}
\description{
Load data into a \code{mipmapper_project} prior to analysis. See
  details for the required data format.
}
\details{
Data must be read in in four parts. Although this may seem 
  cumbersome, it provides an efficient way of storing sample-locus data 
  without duplication of elements (as tends to be found in "long" format).
  This reduces the amount of memory required to store the data.
\itemize{
      \item{\code{data_coverage}}{ a matrix giving the coverage (i.e. the
      total number of reads) for each sample-locus combination. Samples are
      in rows and loci in columns }
      
      \item{\code{data_counts}}{ a matrix giving the barcode counts (i.e. the
      number of reads matched to the Ref allele) for each sample-locus
      combination }
      
      \item{\code{data_sample}}{ a dataframe giving details of each sample. 
      This must have the same number of rows as \code{data_coverage}, and may
      have any number of columns as long as it contains the column
      "Sample_ID" with a unique ID for each sample }
      
      \item{\code{data_loci}}{ a dataframe giving details of each locus. This
      must have the same number of rows as \code{data_coverage} has columns, 
      and may have any number of columns as long as it contains the columns 
      "Chrom" for the chromosome, "Pos" for the genomic position on the 
      chromosome, "Ref" for the reference allele, and "Alt" for the
      alternative allele }
      }
}
