% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{calculate_IBS}
\alias{calculate_IBS}
\title{Calculate proportion identical by state (IBS)}
\usage{
calculate_IBS(x, report_progress = FALSE)
}
\arguments{
\item{x}{a matrix with samples in rows and loci in columns. The value in each
cell should be 1, 0, NA for missing data}

\item{report_progress}{this analysis can take a long time for large datasets,
therefore this option prints running progress to the console}
}
\description{
Calculate the proportion of identity by state (IBS) between all
  pairwise samples. IBS is calculated as the proportion of sites that are
  identical.
}
