% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dummy_data}
\alias{dummy_data}
\title{Dummy data}
\usage{
dummy_data()
}
\value{
Invisibly returns a MIP dataset with the following variables:
  \itemize{
      \item{"Year"}{ Numerics of sample year collect }
      \item{"Country"}{ String of country of sample collection }
      \item{"Study"}{ String of study name }
      \item{"DHS_Cluster"}{ Numeric of DHS cluster}
      \item{"Lat"}{ Numeric for latitude }
      \item{"Long"}{ Numeric for longitude}
      \item{"Sample_ID"}{ String for unique sample ID }
      \item{"Chrom"}{ Numeric for chromosome number }
      \item{"Pos"}{ Numeric for chromosome position}
      \item{"Ref"}{ Reference allele nucleotide}
      \item{"Alt"}{ Alternative allele nucleotide}
      \item{"Coverage"}{ Numeric for read coverage}
      \item{"Barcode_Count"}{ Numeric for count of barcodes matched to Ref}
      }
}
\description{
Create a dummy data set
}
\details{
Creat than a given number of unique reads (coverage).
}
\examples{
dat <- dummy_data()
}
