% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{estimate_f}
\alias{estimate_f}
\title{Estimate pairwise inbreeding coefficient F by maximum likelihood}
\usage{
estimate_f(x, f_breaks = 11, report_progress = FALSE)
}
\arguments{
\item{x}{a matrix with samples in rows and loci in columns. The value in each
cell should be 1, 0, NA for missing data}

\item{f_breaks}{the number of values of f that are explored, distributed
evenly in the [0,1] interval}

\item{report_progress}{this analysis can take a long time for large datasets,
therefore this option prints running progress to the console}
}
\description{
The probability of seeing the same or different alleles at a 
  locus can be written in terms of the global allele frequency p and the 
  inbreeding coefficient f. For example, the probability of seeing the same 
  REF allele is \code{(1-f)*p^2 + f*p}. This formula can be multiplied over 
  all loci to arrive at the overall likelihood of each value of f, which can 
  then be chosen by maximum likelihood. This function carries out this
  comparison between all pairwise samples, passed in as a matrix.
}
