% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fast_read}
\alias{fast_read}
\title{Quickly load data as dataframe}
\usage{
fast_read(file)
}
\arguments{
\item{file}{path to where your data is saved}
}
\description{
Quickly load data as a dataframe.
}
\examples{
\dontrun{
file <- "mip_files/dataset.csv"
dat <- fast_read(file)
}
}
