% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_misc}
\alias{filter_misc}
\title{Filter miscellaneous values}
\usage{
filter_misc(project, filter_SNP = TRUE, integer_counts = TRUE,
  coverage_greq_counts = TRUE, check_delete_output = TRUE)
}
\arguments{
\item{project}{a \code{mipmapper_project} with data already loaded}

\item{filter_SNP}{filter to single nucleotide polymorphism (SNP) only. Remove
all loci where Ref or Alt allele is longer than one character}

\item{integer_counts}{re-code as missing all genotype calls where coverage or
barcode counts are non-integer}

\item{coverage_greq_counts}{re-code as missing all genotype calls where 
coverage is not greater than or equal to barcode counts}

\item{check_delete_output}{whether to prompt the user before overwriting 
existing data/outputs}
}
\description{
Apply miscellaneous filters to data.
}
