% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_dummy_data}
\alias{generate_dummy_data}
\title{Generate Dummy Data}
\usage{
generate_dummy_data(n = 1000, countries = 4, studies = 3, samples = 50,
  pos = 20)
}
\arguments{
\item{n}{Number of rows. Default = 1000}

\item{countries}{Number of different countries. Default = 4}

\item{studies}{Number of different studies. Default = 3}

\item{samples}{Number of different samples. Default = 50}

\item{pos}{Number of different chromosome positions. Default = 20}
}
\value{
Invisibly returns a MIP dataset with the following variables:
  \itemize{
      \item{"Year"}{ Numerics of sample year collect }
      \item{"Country"}{ String of country of sample collection }
      \item{"Study"}{ String of study name }
      \item{"DHS_Cluster"}{ Numeric of DHS cluster}
      \item{"Lat"}{ Numeric for latitude }
      \item{"Long"}{ Numeric for longitude}
      \item{"Sample_ID"}{ String for unique sample ID }
      \item{"Chrom"}{ Numeric for chromosome number }
      \item{"Pos"}{ Numeric for chromosome position}
      \item{"Ref"}{ Reference allele nucleotide}
      \item{"Alt"}{ Alternative allele nucleotide}
      \item{"Coverage"}{ Numeric for read coverage}
      \item{"Barcode_Count"}{ Numeric for count of barcodes matched to Ref}
      }
}
\description{
Create a dummy data set, with control over the creation.
}
\details{
Creates a dummy data set, with more control over how the data
  set variables are created
}
\examples{
dat <- mipmapper:::generate_dummy_data()
}
