% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipmapper_project.R
\name{is.mipmapper_project}
\alias{is.mipmapper_project}
\title{Determine if object is of class mipmapper_project}
\usage{
is.mipmapper_project(x)
}
\arguments{
\item{x}{object of class \code{mipmapper_project}}
}
\description{
Determine if object is of class \code{mipmapper_project}.
}
