% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{latlon_to_bearing}
\alias{latlon_to_bearing}
\title{latlon_to_bearing}
\usage{
latlon_to_bearing(origin_lat, origin_lon, dest_lat, dest_lon)
}
\arguments{
\item{origin_lat}{The origin latitude}

\item{origin_lon}{The origin longitude}

\item{dest_lat}{The destination latitude}

\item{dest_lon}{The destination longitude}
}
\description{
Calculate distance between latitude/longitude points
}
