% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{melt_mip_data}
\alias{melt_mip_data}
\title{Melt mip data}
\usage{
melt_mip_data(data)
}
\arguments{
\item{data}{MIP data. The data must have the following variables:
\itemize{
    \item Chrom : The chromosome number of the MIP read
    \item Pos : The chromosome position of the MIP read
    \item Sample_ID : The name of the sample
    \item Coverage : The total read coverage as numerics
    \item Barcode_Count : The total barocdes recovered
    }}
}
\description{
Melt the mip data into a data frame with one column per locus. 
  The read depth for each locus, denoted by `chr<x>_<pos>`, is given for 
  each sample. An NA is given at loci where there is no read depth.
}
\examples{

data <- data.frame(
"Sample_ID" = c(rep("a", 3), rep("b", 2)),
"Chrom" = c(1, 1, 2, 1, 1),
"Pos" = c(100, 200, 50, 100, 200),
"Coverage" = c(47, 95, 100, 52, 74),
"Barcode_Count" = c(47, 0, 40, 52, 70)
)

melt_mip_data(data)

}
