% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipmapper.R
\docType{package}
\name{mipmapper}
\alias{mipmapper}
\alias{mipmapper-package}
\alias{mipmapper-package}
\title{\pkg{mipmapper} Filtering and analysis of MIP panel genetic datasets}
\description{
Molecular inversion probes (MIPs) are a tool for genomic
partitioning, which result in datasets of SNP genotypes. mipmapper
is a package to help assist in filtering the raw output from these
sequencing efforts, and conducting population genetic analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mrc-ide.github.io/mipmapper/}
  \item Report bugs at \url{https://github.com/mrc-ide/mipmapper/issues}
}

}
\author{
\strong{Maintainer}: Bob Verity \email{r.verity@imperial.ac.uk}

Authors:
\itemize{
  \item OJ Watson
}

}
