% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mipmapper_file}
\alias{mipmapper_file}
\title{Load system file}
\usage{
mipmapper_file(name)
}
\arguments{
\item{name}{the file name}
}
\description{
Load system file
}
\details{
Load a file from within the mipmapper package.
}
\examples{
# TODO
}
