% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{perform_pca}
\alias{perform_pca}
\title{Perform PCA analysis}
\usage{
perform_pca(project)
}
\arguments{
\item{project}{a \code{mipmapper_project} with data already loaded}
}
\description{
Principal component analysis will be condcuted and the resultant
  components returned, along with the variance explained by each component
  and the loadings of each component.
}
