% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot coverage gaps in data}
\usage{
plot_coverage(dat, type = "o", pch = 20, ylim = c(0, 100),
  xlab = "Coverage (unique reads)", ylab = "data completeness (\%)",
  log = "x", ...)
}
\arguments{
\item{dat}{TODO}

\item{type}{TODO}

\item{pch}{TODO}

\item{ylim}{TODO}

\item{xlab}{TODO}

\item{ylab}{TODO}

\item{log}{TODO}

\item{...}{other arguments to \code{plot()}}
}
\description{
Plot coverage gaps in data
}
\details{
TODO
}
\examples{
# TODO
}
