% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_coverage_matrix}
\alias{plot_coverage_matrix}
\title{Plot raster image of coverage matrix}
\usage{
plot_coverage_matrix(project)
}
\arguments{
\item{project}{a \code{mipmapper_project} with data already loaded}
}
\description{
Plot raster image of coverage matrix.
}
