% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_DRC}
\alias{plot_DRC}
\title{Plot points on map of DRC}
\usage{
plot_DRC(longitude = NULL, latitude = NULL, v = NULL,
  longitude_limits = c(12, 33), latitude_limits = c(-15, 7), vmin = min(v,
  na.rm = TRUE), vmax = max(v, na.rm = TRUE), plot_water = TRUE,
  raw_cols = viridis(100))
}
\arguments{
\item{longitude}{x-coordinates of points to plot}

\item{latitude}{y-coordinates of points to plot}

\item{v}{plotting value}

\item{longitude_limits}{plotting x-limits}

\item{latitude_limits}{plotting y-limits}

\item{vmin}{minimum value used in colour scale}

\item{vmax}{maximum value used in colour scale}

\item{plot_water}{whether to overlay major water bodies}

\item{raw_cols}{the cols that make up the colour palette}
}
\description{
Plot points on map of DRC
}
