% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA}
\usage{
plot_pca(project, num_components = 2, grouping_var = NULL)
}
\arguments{
\item{project}{a \code{mipmapper_project} containing pca analysis output}

\item{num_components}{numeric for number of components used}

\item{grouping_var}{character for the desired meta variable to be used for 
labelling the scatterplot}
}
\description{
Plots either the first 2 or 3 principal components, with the
  data points labelled accordng to chosen meta data.
}
\examples{
\dontrun{
dat <- dummy_data()
dat <- filter_misc(dat = dat)
dat <- filter_coverage(dat = dat, min_coverage = 2)
dat <- melt_mip_data(dat = dat)
dat <- impute_mip_data(dat = dat)
pca <- pca_mip_data(dat = dat)
plot_pca(pca, num_components = 2, meta_var = "Country")
plot_pca(pca, num_components = 3, meta_var = "Study")
}
}
