% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pca_loadings}
\alias{plot_pca_loadings}
\title{Plot loading values of PCA}
\usage{
plot_pca_loadings(project, component = 1, colour = NULL,
  plot_resistance = TRUE)
}
\arguments{
\item{project}{a \code{mipmapper_project} containing pca analysis output}

\item{component}{which principal component to plot}

\item{colour}{optional name of variable to colour by}

\item{plot_resistance}{whether to overlay known drug resistance loci}
}
\description{
Plot loading values of PCA for a given component.
}
