% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_WSAF_matrix}
\alias{plot_WSAF_matrix}
\title{Plot raster image of within-sample allele frequencies}
\usage{
plot_WSAF_matrix(project)
}
\arguments{
\item{project}{a \code{mipmapper_project} with data already loaded}
}
\description{
Plot raster image of within-sample allele frequencies.
}
