% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipmapper_project.R
\name{print.mipmapper_project}
\alias{print.mipmapper_project}
\title{Custom print function for class mipmapper_project}
\usage{
\method{print}{mipmapper_project}(x, ...)
}
\arguments{
\item{x}{object of class \code{mipmapper_project}}

\item{...}{other arguments (ignored)}
}
\description{
Custom print function for class \code{mipmapper_project},
  printing a summary of the key elements (also equivalent to
  \code{summary(x)}). To do an ordinary \code{print()} of all elements of the
  project, use the \code{print_full()} function.
}
