% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipmapper_project.R
\name{summary.mipmapper_project}
\alias{summary.mipmapper_project}
\title{Print summary for class mipmapper_project}
\usage{
summary.mipmapper_project(object, ...)
}
\arguments{
\item{object}{object of class \code{mipmapper_project}}

\item{...}{other arguments (ignored)}
}
\description{
Custom summary function for class \code{mipmapper_project}.
}
