% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{fast_read}
\alias{fast_read}
\title{Quickly load data as data.frame}
\usage{
fast_read(file)
}
\arguments{
\item{file}{String to the path where your mip dataset is saved}
}
\description{
Quickly load data as data.frame
}
\details{
TODO
}
\examples{
\dontrun{
file <- "mip_files/dataset.csv"
dat <- fast_read(file)
}
}
