% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_coverage}
\alias{filter_coverage}
\title{Filter data}
\usage{
filter_coverage(dat, min_coverage = 2)
}
\arguments{
\item{dat}{MIP data. The data must have the following variables:
\itemize{
    \item Coverage : The total read coverage as numerics
    }}

\item{min_coverage}{Numeric for the minimum required coverage. Default = 2}
}
\description{
Filter data based on coverage
}
\details{
The MIP dataset is filtered to remove all loci that have less than 
  the minimum required coverage.
}
\examples{
dat <- dummy_data()
dat <- filter_misc(dat)
dat2 <- filter_coverage(dat, min_coverage = 5)
dim(dat2)[1] < dim(dat)[1]
}
