% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_coverage}
\alias{filter_coverage}
\title{Filter data}
\usage{
filter_coverage(dat, min_coverage = 2)
}
\arguments{
\item{dat}{TODO}

\item{min_coverage}{TODO}
}
\description{
Filter data based on coverage
}
\details{
TODO
}
\examples{
# TODO
}
