% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{filter_misc}
\alias{filter_misc}
\title{Filter data}
\usage{
filter_misc(dat, SNP_only = TRUE, group_Alt = TRUE, drop_irregular = TRUE)
}
\arguments{
\item{dat}{TODO}

\item{SNP_only}{TODO}

\item{group_Alt}{TODO}

\item{drop_irregular}{TODO}
}
\description{
Filter out miscellaneous unwanted values from data
}
\details{
TODO
}
\examples{
# TODO
}
