% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{impute}
\alias{impute}
\title{Impute missing values}
\usage{
impute(dat, MARGIN = 2, FUN = mean, ...)
}
\arguments{
\item{dat}{a matrix of values}

\item{MARGIN}{which margin to apply summary function over: 1 = rows, 2 = 
columns}

\item{FUN}{function to impute missing values. Default = `mean`}

\item{...}{other arguments to pass to FUN}
}
\description{
Missing values in the input matrix are imputed by applying a 
  summary function over rows or columns. The summary function is only applied to
  non-NA values. The default summary function takes the mean.
}
