% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{pca_mip_data}
\alias{pca_mip_data}
\title{PCA of MIP Data}
\usage{
pca_mip_data(project, plot_on = TRUE)
}
\arguments{
\item{project}{a \code{mipmapper_project} with data already loaded}

\item{plot_on}{whether to plot PCA results}
}
\description{
Using the output of \code{impute_mip_data}, principal component
  analysis will be condcuted and the resultant components returned, along
  with the variance explained by each component and the loadings of each
  component.
}
