% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pca_variance}
\alias{plot_pca_variance}
\title{Plot variance explained by PCA components}
\usage{
plot_pca_variance(project, num_components = 10)
}
\arguments{
\item{project}{a \code{mipmapper_project} containing pca analysis output}

\item{num_components}{number of components to be shown}
}
\description{
Plot the variance explained by each component. The number of 
  components shown is controlled by \code{num_components}, with up to the
  first 10 components shown by default.
}
